workflow_manager/requirements.txt
Flask
paramiko
prometheus-flask-exporter
PyJWT
pymongo
python-dotenv
pyyaml